config pkg; move DiscrodBotToken to config;
