+++ cmd/trash/main.go
@@ -1,30 +1,28 @@
 package main
 
 import (
-	"errors"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"os/signal"
 	"strings"
 	"syscall"
 	"time"
 
+	"github.com/KKopilka/discord-bot/internal/config"
 	"github.com/KKopilka/discord-bot/internal/service"
 	"github.com/bwmarrin/discordgo"
 )
-
-var discordBotToken string
 
 const WasteBasketEmoji = "🗑"
 
 func main() {
 	// 1. Загрузка конфигурации пакет config
-	err := readBotToken()
-	if err != nil {
+	if err := config.ReadConfig(); err != nil {
 		fmt.Println(err.Error())
 		return
 	}
+
+	discordBotToken := config.BotToken()
 
 	fmt.Println("Configuration readed successfully. Create and start bot service.", discordBotToken, "lol")
 	// 2. Структура сервиса бота пакет service
@@ -154,53 +152,6 @@ func IsChannelNeedToClean(channelName string) bool {
 		return true
 	}
 	return false
-}
-
-type DiscordBotToken string
-
-func (bt *DiscordBotToken) Sanitize() {
-	content := string(*bt)
-	content = strings.Trim(content, " \n\r\t")
-	// TODO: check bot-token specification for whitespaces inside it.
-	content = strings.ReplaceAll(content, " ", "")
-
-	*bt = DiscordBotToken(content)
-}
-
-func (bt DiscordBotToken) Validate() error {
-	content := string(bt)
-
-	if strings.Contains(content, " ") {
-		return errors.New("whitespace in bot-token")
-	}
-	if strings.Contains(content, "\n") {
-		return errors.New("end of line in bot-token")
-	}
-
-	if len(content) != 72 {
-		return errors.New(fmt.Sprintln("invalid bot-token. Token length:", len(content)))
-	}
-
-	return nil
-}
-
-func readBotToken() error {
-	file, err := ioutil.ReadFile("bot-token")
-
-	if err != nil {
-		return err
-	}
-
-	botToken := DiscordBotToken(string(file))
-	botToken.Sanitize()
-
-	if err := botToken.Validate(); err != nil {
-		return err
-	}
-
-	discordBotToken = string(botToken)
-
-	return nil
 }
 
 func processGuilds(s *service.Service) error {
